Allow restoring use_global flag in set_privacy
I toggled the privacy mode for my camera on and then off again in Home Assistant. I was mystified to find that it subsequently stopped recording when it had been recording before.

Turns out the camera previously had `use_global` set to `True`, so that it used the global settings for recording schedule. Changing its recording mode to `NEVER` disabled `use_global`, and it was not correctly restored later.

Add a parameter that will allow a PR in Home Assistant to restore the previous `use_global` flag.